use const int in array declaration

+++ src/lines-as-network.cpp
@@ -73,7 +73,7 @@ Rcpp::List rcpp_lines_as_network (const Rcpp::List &sf_lines)
     Rcpp::CharacterVector owb = sf_lines [oneWayBicycleIndex];
 
     Rcpp::List geoms = sf_lines [nms.size () - 1];
-    bool isOneWay [geoms.length ()] = {false};
+    bool isOneWay [(const int) geoms.length ()] = {false};
     // Get dimension of matrix
     size_t nrows = 0;
     int ngeoms = 0;
